Problem: communication with Rust WebAssembly modules
Implementing long-running processes in Rust WebAssembly modules is a bit
problematic. We can't run them as is on the main thread because they
will block it and that's not great.

However, simply moving them to run in the worker only works if we never
need to communicate with it. It can't really observe worker messages as
this requires giving control back to JavaScript environment.

Solution: implement a communication channel on top of `SharedArrayBuffer`

Please note that the algorithm used in `spsc` channel has not been
extensively vetted and is subject to change.

I would also like to note that target application deployment is still a
bit cumbersome as we have to generate *two* wasm-bindgen targets:
bundler and no-modules in order for workers to be able to use it. This
is not ideal but can be address later (hopefully)